Fix fromdate issue
Closes #1

This was just a dumb mistake, from setting lastTime just before the
request is made.
Setting the timestamp after the request is in line with the original
Boson's inner workings. See
https://github.com/SOBotics/Boson/blob/af86f295ab622867709c9f2ef3d1092faa947225/src/main/java/org.sobotics.boson/framework/services/chat/monitors/CommentMonitor.java#L76

Not sure if this can cause race conditions in the other direction (i.e.
where comments are discarded because rounding and shit), but I strongly
doubt it. This would be extremely specific circumstances as well.